Componentize modal
Componentize modal. Modal can be used as a component. Clicks outside of the modal and esc key pressing dismisses the modal. Body is whatever is passed into the child slot.

Additionally, improve small screen layout with the filter options. For now, we just remove them for small screens, but we could implement a menu or something for small screens.